fix(Calculator): error handler
Prior to this commit, when user get an error, user can continue do the
calculation with `Infinity` and `null`. So I disabled all of the buttons
except "C" if user get an error.

On the other hand, the error message "Cannot devide by zero" is too long
to show in our ui. I increase the width of the calculator to avoid this.